+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 17.3.0:
 
 - Added support for 1.20 (no changes required, already compatible).
+- Fixed quit messages not working with AuthMe.
 
 ## 17.2.4: